requirements.txt
# Core
discord.py>=2.0.0  # Discord bot library

# Database
sqlalchemy>=1.4.0  # ORM for MySQL
pymysql>=1.0.0     # MySQL driver

# Environment
python-dotenv>=0.19.0  # Load .env variables